refactor date helpers
