+++ src/helpers/dateHelper.ts
@@ -8,3 +8,62 @@ export const getExpirationDate = (): string => {
   res.setTime(expDate);
   return res.toISOString();
 };
+
+export const getNext1pmOr1am = (): Date => {
+  const nextSyncDate = new Date();
+  if (nextSyncDate.getHours() < 13) {
+    nextSyncDate.setHours(13); // same day at 1pm
+    nextSyncDate.setMinutes(0);
+    nextSyncDate.setSeconds(0);
+    nextSyncDate.setMilliseconds(0);
+  } else {
+    nextSyncDate.setTime(Date.now() + 24 * 3600 * 1000); // next day same hour
+    nextSyncDate.setHours(1); // at 1 am
+    nextSyncDate.setHours(0);
+    nextSyncDate.setMinutes(0);
+    nextSyncDate.setSeconds(0);
+    nextSyncDate.setMilliseconds(0);
+  }
+  return nextSyncDate;
+};
+
+export const getNext8am = (): Date => {
+  const notificationDate = new Date();
+  // first get next 8am time
+  if (notificationDate.getHours() < 8) {
+    notificationDate.setHours(8); // same day at 8am
+    notificationDate.setMinutes(0);
+    notificationDate.setSeconds(0);
+    notificationDate.setMilliseconds(0);
+  } else {
+    notificationDate.setTime(notificationDate.getTime() + 24 * 3600 * 1000); // next day same hour
+    notificationDate.setHours(8); // at 8 am
+    notificationDate.setMinutes(0);
+    notificationDate.setSeconds(0);
+    notificationDate.setMilliseconds(0);
+  }
+
+  // then allow only mondays, wednesdays and fridays
+  const d = notificationDate.getDay();
+  if (d == 0 || d == 2 || d == 4) {
+    // sunday -> monday
+    // tuesday -> wednesday
+    // thursday -> friday
+    notificationDate.setTime(notificationDate.getTime() + 24 * 3600 * 1000);
+  } else if (d == 6) {
+    // saturday -> monday
+    notificationDate.setTime(notificationDate.getTime() + 48 * 3600 * 1000);
+  }
+  return notificationDate;
+};
+
+export const getNextMidnight = (): Date => {
+  const nextSyncDate = new Date();
+  nextSyncDate.setTime(Date.now() + 24 * 3600 * 1000); // next day same hour
+  nextSyncDate.setHours(0);
+  nextSyncDate.setHours(0);
+  nextSyncDate.setMinutes(0);
+  nextSyncDate.setSeconds(0);
+  nextSyncDate.setMilliseconds(0);
+  return nextSyncDate;
+};
